Move DerivationFormatter component into its own file
